refactor(harvest): backend 子包抽取，harvest.py 2451→1632 行零循环 DAG (v1.11.0) (#101)
将巨型 harvest.py 按单向无环依赖 DAG 拆为主编排层 + 三子模块，纯搬迁无外部行为变更：

- harvest_safety.py：SSRF 守卫 / 域黑名单 / 路径沙箱 / 限流 / URL 规范化（零依赖叶子）
- harvest_search/：4 个 web search backend + do_search 编排
- harvest_fetch/：4 个 URL fetch backend + do_fetch 编排
- harvest_clients/base.py：curl_cffi 传输能力 + fail-fast 检查随 _fetch_curl_cffi 上移
- test_harvest.py：内部符号引用改走子模块（facade-penetration mock），438 用例不变全绿
- 新增插件级 CLAUDE.md 代码地图
- 双写版本 1.10.4 → 1.11.0
